refactor: rename MySKPageHeader file name
